backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Password hashing
argon2-cffi==23.1.0

# Environment and config
python-dotenv==1.0.0

# Utilities
email-validator==2.1.0

# Production server
gunicorn==21.2.0
